misc: regroup download file methods into one util
