+++ src/components/customerPortal/PortalInvoicesList.tsx
@@ -17,9 +17,9 @@ import {
   Typography,
 } from '~/components/designSystem'
 import { SearchInput } from '~/components/SearchInput'
-import { addToast } from '~/core/apolloClient'
 import { intlFormatNumber } from '~/core/formats/intlFormatNumber'
 import { deserializeAmount } from '~/core/serializers/serializeAmount'
+import { handleDownloadFile } from '~/core/utils/downloadFiles'
 import {
   CurrencyEnum,
   InvoiceForFinalizeInvoiceFragmentDoc,
@@ -203,26 +203,7 @@ const PortalInvoicesList = () => {
 
   const [downloadInvoice] = useDownloadCustomerPortalInvoiceMutation({
     onCompleted(localData) {
-      const fileUrl = localData?.downloadCustomerPortalInvoice?.fileUrl
-
-      if (fileUrl) {
-        // We open a window, add url then focus on different lines, in order to prevent browsers to block page opening
-        // It could be seen as unexpected popup as not immediatly done on user action
-        // https://stackoverflow.com/questions/2587677/avoid-browser-popup-blockers
-        const myWindow = window.open('', '_blank')
-
-        if (myWindow?.location?.href) {
-          myWindow.location.href = fileUrl
-          return myWindow?.focus()
-        }
-
-        myWindow?.close()
-      } else {
-        addToast({
-          severity: 'danger',
-          translateKey: 'text_62b31e1f6a5b8b1b745ece48',
-        })
-      }
+      handleDownloadFile(localData?.downloadCustomerPortalInvoice?.fileUrl)
     },
   })
