+++ src/components/invoices/InvoicesList.tsx
@@ -26,6 +26,7 @@ import { CUSTOMER_INVOICE_DETAILS_ROUTE, INVOICE_SETTINGS_ROUTE } from '~/core/r
 import { deserializeAmount } from '~/core/serializers/serializeAmount'
 import { formatDateToTZ } from '~/core/timezone'
 import { copyToClipboard } from '~/core/utils/copyToClipboard'
+import { handleDownloadFile } from '~/core/utils/downloadFiles'
 import {
   CurrencyEnum,
   GetInvoicesListQuery,
@@ -78,26 +79,7 @@ const InvoicesList = ({
 
   const [downloadInvoice] = useDownloadInvoiceItemMutation({
     onCompleted({ downloadInvoice: data }) {
-      const fileUrl = data?.fileUrl
-
-      if (fileUrl) {
-        // We open a window, add url then focus on different lines, in order to prevent browsers to block page opening
-        // It could be seen as unexpected popup as not immediatly done on user action
-        // https://stackoverflow.com/questions/2587677/avoid-browser-popup-blockers
-        const myWindow = window.open('', '_blank')
-
-        if (myWindow?.location?.href) {
-          myWindow.location.href = fileUrl
-          return myWindow?.focus()
-        }
-
-        myWindow?.close()
-      } else {
-        addToast({
-          severity: 'danger',
-          translateKey: 'text_62b31e1f6a5b8b1b745ece48',
-        })
-      }
+      handleDownloadFile(data?.fileUrl)
     },
   })
